+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### 0.1.3: db options / space fix
+- fixed an issue where a space was often appended to the start of strings
+- added options for only logging to a database or only logging to a console, despite possibly being provided database details - not yet fully implemented
+- added a 'test' script to `../package.json`
+
 ### 0.1.2: brighter colors
 - made most colors a little brighter
 - added a test for generic logging